feat: translation fields
