+++ src/hooks.server.ts
@@ -1,0 +1,12 @@
+import type { Handle } from '@sveltejs/kit'
+import { locale } from 'svelte-i18n'
+
+export const handle: Handle = async ({ event, resolve }) => {
+	const lang = event.request.headers.get('accept-language')?.split(',')[0]
+	if (lang) {
+		locale.set(lang)
+	} else {
+        locale.set('en')
+    }
+	return resolve(event)
+}
