+++ src/routes/lang/eo.json
