--- src/routes/login/+layout.svelte
@@ -1,5 +1,0 @@
-<script>
-    import '../../styles/login.css'
-</script>
-
-<slot />
